requirements.txt
django>=1.5.1
wheel==0.23.0
djangorestframework>=2.3.14
